Move module into management command
